# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Canon Camera Platform API
  version: 0.1.0
  summary: Edge and control-plane APIs for controlling Canon cameras
  description: |
    This contract documents both sides of the first Canon camera automation platform:
    - the edge API running next to the camera over USB
    - the control API orchestrating one or more trusted edge devices

    Initial hardware target: Canon EOS R50 via `gphoto2` / `libgphoto2`.
servers:
  - url: https://edge-camera.local
    description: Example private edge endpoint
  - url: https://control-camera.local
    description: Example private control-plane endpoint
security:
  - bearerAuth: []
  - mutualTLS: []
...
paths:
...
  /v1/profiles:
    get:
      tags: [Profiles]
      summary: List saved camera profiles
      operationId: listProfiles
      responses:
        '200':
          description: Camera profiles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileListResponse'
    post:
      tags: [Profiles]
      summary: Create a saved camera profile
      operationId: createProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileRequest'
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CameraProfile'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/profiles/{profileId}:
    get:
      tags: [Profiles]
      summary: Get a saved camera profile
      operationId: getProfile
      parameters:
        - $ref: '#/components/parameters/ProfileId'
      responses:
        '200':
          description: Saved profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CameraProfile'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Profiles]
      summary: Update a saved camera profile
      operationId: updateProfile
      parameters:
        - $ref: '#/components/parameters/ProfileId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CameraProfile'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Profiles]
      summary: Delete a saved camera profile
      operationId: deleteProfile
      parameters:
        - $ref: '#/components/parameters/ProfileId'
      responses:
        '204':
          description: Profile deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/profiles/{profileId}/apply:
    post:
      tags: [Profiles]
      summary: Apply a saved camera profile to the attached camera
      operationId: applyProfile
      parameters:
        - $ref: '#/components/parameters/ProfileId'
        - $ref: '#/components/parameters/SessionTokenHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      responses:
        '202':
          description: Profile apply job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedJobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/sessions:
    post:
      tags: [Sessions]
      summary: Create an exclusive camera control session
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '409':
          description: Camera already leased by another owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/sessions/{sessionId}:
    get:
      tags: [Sessions]
      summary: Get session details
      operationId: getSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Sessions]
      summary: Release a camera control session
      operationId: releaseSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/SessionTokenHeader'
      responses:
        '204':
          description: Session released
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/sessions/{sessionId}/renew:
    post:
      tags: [Sessions]
      summary: Extend an active session's lease
      description: A lease is fixed at creation and is not refreshed by activity; long-lived clients (e.g. continuous preview polling) must call this periodically to avoid the session expiring mid-use.
      operationId: renewSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/SessionTokenHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                leaseSeconds:
                  type: integer
                  minimum: 15
                  maximum: 3600
                  default: 120
      responses:
        '200':
          description: Session renewed with a new expiry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Session is no longer active and cannot be renewed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/camera/configs/{key}:
    patch:
      tags: [Camera]
      summary: Update a normalized camera config
      operationId: updateCameraConfig
      parameters:
        - $ref: '#/components/parameters/ConfigKey'
        - $ref: '#/components/parameters/SessionTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCameraConfigRequest'
      responses:
        '200':
          description: Updated config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CameraConfig'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/camera/preview:
    get:
      tags: [Camera]
      summary: Capture a current preview frame
      operationId: getCameraPreview
      parameters:
        - $ref: '#/components/parameters/SessionTokenHeader'
      responses:
        '200':
          description: Preview frame
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          description: Preview not available for current camera state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/camera/focus/autofocus:
    post:
      tags: [Focus]
      summary: Trigger autofocus
      operationId: triggerAutofocus
      parameters:
        - $ref: '#/components/parameters/SessionTokenHeader'
      responses:
        '202':
          description: Autofocus accepted as a job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedJobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/camera/focus/steps:
    post:
      tags: [Focus]
      summary: Move focus by a relative number of steps
      operationId: stepFocus
      parameters:
        - $ref: '#/components/parameters/SessionTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FocusStepRequest'
      responses:
        '202':
          description: Focus step accepted as a job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedJobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/captures:
    post:
      tags: [Capture]
      summary: Trigger a still capture job
      operationId: createCapture
      parameters:
        - $ref: '#/components/parameters/SessionTokenHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaptureRequest'
      responses:
        '202':
          description: Capture job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedJobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /v1/storage/files:
    get:
      tags: [Storage]
      summary: List files from a camera folder
      operationId: listStorageFiles
      parameters:
        - name: volumeId
          in: query
          required: true
          schema:
            type: string
        - name: folderPath
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Files in the requested folder
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageFileListResponse'
  /v1/storage/files/download:
    post:
      tags: [Storage]
      summary: Download a file from the camera to edge storage
      operationId: downloadStorageFile
      parameters:
        - $ref: '#/components/parameters/SessionTokenHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadStorageFileRequest'
      responses:
        '202':
          description: Download job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedJobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/storage/files/delete:
    post:
      tags: [Storage]
      summary: Delete a file from camera storage
      operationId: deleteStorageFile
      parameters:
        - $ref: '#/components/parameters/SessionTokenHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteStorageFileRequest'
      responses:
        '202':
          description: Delete job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedJobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/jobs/{jobId}:
    get:
      tags: [Jobs]
      summary: Get a job by ID
      operationId: getJob
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/media/{assetId}:
    get:
      tags: [Media]
      summary: Get metadata for a locally stored media asset
      operationId: getMediaAsset
      parameters:
        - $ref: '#/components/parameters/AssetId'
      responses:
        '200':
          description: Media asset metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaAsset'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/media/{assetId}/content:
    get:
      tags: [Media]
      summary: Download the raw bytes of a locally stored media asset
      operationId: getMediaAssetContent
      parameters:
        - $ref: '#/components/parameters/AssetId'
      responses:
        '200':
          description: Media asset file content, served with the asset's stored mimeType
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: Media asset record not found, or its file is no longer present on disk
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    mutualTLS:
      type: mutualTLS
  parameters:
...
    ProfileId:
      name: profileId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SessionTokenHeader:
      name: X-Session-Token
      in: header
      required: true
      schema:
        type: string
      description: Lease token returned when the session was created
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Client-supplied key for safely retrying mutating operations
  responses:
    Unauthorized:
      description: Request is missing or using an invalid session/auth token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with current camera or session state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ProfileListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CameraProfile'
    CreateProfileRequest:
      type: object
      required: [name, settings]
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        settings:
          $ref: '#/components/schemas/ProfileSettings'
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        settings:
          $ref: '#/components/schemas/ProfileSettings'
    CameraProfile:
      type: object
      required: [profileId, name, settings, createdAt, updatedAt]
      properties:
        profileId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        settings:
          $ref: '#/components/schemas/ProfileSettings'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastAppliedAt:
          type: string
          format: date-time
          nullable: true
...